% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmisd_functions.R
\name{br_plot}
\alias{br_plot}
\title{Background Rate Plot}
\usage{
br_plot(model, colors = c("blue", "red"), plot_title = NULL)
}
\arguments{
\item{model}{is object exported from ctmisd function}

\item{colors}{is a vector of two colors representing low and high values}

\item{plot_title}{optional title to include}
}
\value{
a ggplot object showing the nonstationary background rate.
}
\description{
Returns a heat map showing the nonstationary background rate over the spatial domain
}
