% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmisd_functions.R
\name{cond_int_smooth}
\alias{cond_int_smooth}
\title{Conditional Intensity}
\usage{
cond_int_smooth(model)
}
\arguments{
\item{model}{the output from the ctmisd function.}
}
\value{
a data frame containing the time, date, latitude, longitude, mark, and
conditional intensity at each observed point.
}
\description{
This function calculates the conditional intensity using CTMISD.
}
