% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmisd_functions.R
\name{ctmisd}
\alias{ctmisd}
\title{Model Independent Stochastic Declustering}
\usage{
ctmisd(
  dates,
  ref_date = 0,
  lat = NA,
  lon = NA,
  lon_lim = NA,
  lat_lim = NA,
  marks = NA,
  stopwhen = 0.001,
  time_of_day = NA,
  just_times = FALSE,
  time_unit = "day",
  dist_unit = "mile",
  n_ints = 1000,
  g_span = 0.1,
  h_span = 0.1,
  k_span = 0.1,
  nonstat_br = TRUE,
  dist = c("beta", 0.5, 1),
  np = 20,
  time_len = 0.01,
  space_len = 0.01,
  show_progress = FALSE
)
}
\arguments{
\item{dates}{a vector of dates as "yyyy-mm-dd".}

\item{ref_date}{a date to serve as time 0, defaults to earliest observation}

\item{lat}{a vector of latitudes, omit if not using spatial data}

\item{lon}{a vector of longitudes, omit if not using spatial data}

\item{lon_lim}{vector of minimum, maximum, and width of pixels for nonstationary background rate}

\item{lat_lim}{vector of minimum, maximum, and height of pixels for nonstationary background rate}

\item{marks}{a vecotr of marks, or magnitudes, omit if not using marked data}

\item{time_of_day}{character string that lists the time of day of events, as hour:minute:second}

\item{just_times}{TRUE or FALSE object. TRUE if \code{dates} object is a vector of only times,
indicating time elapsed since beginning of catalog.}

\item{time_unit}{character string that specifies the desired unit of time}

\item{dist_unit}{character string that specifies the desired unit of distance: meter, kilometer, or mile}

\item{n_ints}{integer defining the number of temporal and spatial intervals to be used
to estimate triggering functions}

\item{g_span}{smoothing parameter (span) of LOESS used in estimating g(t)}

\item{h_span}{smoothing parameter (span) of LOESS used in estimating h(s)}

\item{k_span}{smoothing parameter (span) of LOESS used in estimating k(m)}

\item{nonstat_br}{TRUE if using a nonstationary background rate}

\item{dist}{vector specifying the sampling distribution of intervals and required parameters}

\item{np}{integer defining the number of neighboring points to use for variable bandwidth
of kernel density smoothing for nonstationary background rate}

\item{time_len}{double between 0 and 1 defining the maximum temporal interval length as a proportion
of the temporal domain}

\item{space_len}{double between 0 and 1 defining the maximum spatial interval length as a proportion
of the spatial domain}

\item{show_progress}{when TRUE, algorithm will print the iteration number and maximum
pairwise change in the probability matrix.}

\item{time_breaks}{a vector of cutoff values for temporal bins of time differences}

\item{space_breaks}{a vector of cutoff values for spatial bins of distance differences}

\item{mark_breaks}{a vector of cutoff values for magnitude bins}

\item{stop_when}{scalar that serves as conversion criterion, 1e-3 as default}
}
\value{
Probability matrix \code{p0} containing the probabilities that event
\code{i} is an offspring of event \code{j}, \code{i > j}. Diagonal elements
represent the probability that event \code{i} is a background event.

\code{br} is the estimated background rate of the process

\code{perc_diag} is the proportion of mass lying on the diagonal of matrix \code{p0}

\code{perc_br} is the proportion of events in which the maximum probabilistic assignment
is as a background event

\code{n_iterations} is the number of iterations executed until convergence

\code{locs} is a data frame listing midpoint latitude, midpoint longitude,
x and y index, and background rate of the pixel each event lies in, for nonstationary background rate

\code{br_grid} is a data frame with the background rate and midpoint of each pixel

\code{g_integral} is the Reimann sum used to scale g(t) as a density

\code{h_integral} is the Reimann sum used to scale h(s) as a density

\code{fitg} is a list containing the output of the LOESS fit of g(t)

\code{fith} is a list containing the output of the LOESS fit of h(s)

\code{fitk} is a list containing the output of the LOESS fit of k(m)

\code{input} is a list of all inputs
}
\description{
This function uses nonparametric procedures to analyze a Hawkes
process in temporal or spatio-temporal domain, with or without marks
through the Model independent stochastic declustering algorithm.
}
