% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmisd_functions.R
\name{dev_plot}
\alias{dev_plot}
\title{Deviance residual plot}
\usage{
dev_plot(
  dev1,
  dev2,
  plot_title = NULL,
  obs_data = NULL,
  lon_lim = seq(min(obs_data$lon) - 0.01, max(obs_data$lon) + 0.01, (max(obs_data$lon)
    - min(obs_data$lon))/10),
  lat_lim = seq(min(obs_data$lat) - 0.01, max(obs_data$lat) + 0.01, (max(obs_data$lat)
    - min(obs_data$lat))/10),
  include_pts = TRUE,
  include_vals = FALSE
)
}
\arguments{
\item{dev1}{the deviance residuals object from the first model}

\item{dev2}{the deviance residuals object from the second model}

\item{plot_title}{the optional character string to be used for the plot title}

\item{obs_data}{the observational data used in fitting the models}

\item{lon_lim}{a vector of the minimum longitude, maximum longitude, and width of each pixel}

\item{lat_lim}{a vector of the minimum latitude, maximum latitude, and height of each pixel}

\item{include_pts}{TRUE if observational data is plotted}

\item{include_vals}{TRUE if the deviance residual values are plotted in each pixel}
}
\description{
Plot displaying the deviance residuals over the spatial domain.
}
