% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmisd_functions.R
\name{dev_res}
\alias{dev_res}
\title{Calculates the deviance residuals of a nonparametric Hawkes process model fit
using the misd or ctmisd function. May be used compare the performance of two
competing models.}
\usage{
dev_res(
  model1,
  model2 = NULL,
  time_pts = quantile(1:ceiling(max(model1$data$times)), probs = seq(0, 1, 0.05)),
  bin1 = TRUE
)
}
\arguments{
\item{model1}{the output from the misd or ctmisd function.}

\item{model2}{the output from an optional second model fit with misd or ctmisd}

\item{time_pts}{a vector of points spanning the temporal domain.
The conditional intensity will be evaluated at each point.}

\item{bin1}{TRUE if model 1 was estimated using nphawkes::misd, rather than ctmisd}
}
\value{
a data frame of each pixel's x and y coordinate, latitude and longitude midpoint,
and deviance residual
}
\description{
Calculates the deviance residuals of a nonparametric Hawkes process model fit
using the misd or ctmisd function. May be used compare the performance of two
competing models.
}
