% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctmisd_functions.R
\name{trig_plots}
\alias{trig_plots}
\title{This function plots the smooth triggering function for each component,
g(t), h(s), and k(m) for time, space, and marks, respectively.}
\usage{
trig_plots(
  model,
  g_xlim = c(min(model$fitg$x), max(model$fitg$x)),
  h_xlim = c(min(model$h$midpt), max(model$h$midpt)),
  k_xlim = c(min(model$k$midpt), max(model$k$midpt)),
  g_ylim = c(0, NA),
  h_ylim = c(0, NA),
  k_ylim = c(0, NA),
  mag_label = "magnitude",
  se_include = TRUE,
  include_pts = F
)
}
\arguments{
\item{model}{the output from the ctmisd() function.}

\item{g_xlim}{a vector of the minimum and maximum value on the x-axis of g(t)}

\item{h_xlim}{a vector of the minimum and maximum value on the x-axis of h(s)}

\item{k_xlim}{a vector of the minimum and maximum value on the x-axis of k(m)}

\item{g_ylim}{a vector of the minimum and maximum value on the y-axis of g(t)}

\item{h_ylim}{a vector of the minimum and maximum value on the y-axis of h(s)}

\item{k_ylim}{a vector of the minimum and maximum value on the y-axis of k(m)}

\item{mag_label}{a character for the name of the marks employed}

\item{se_include}{TRUE if standard errors are to be included on the plots}

\item{include_pts}{TRUE if points, in addition to the fitted curves, are to be included on the plots}
}
\value{
\code{all_plots} all triggering component studied are plotted

\code{time_plot} only g(t) is shown

\code{space_plot} only h(s) is shown

\code{mark_plot} only k(m) is shown
}
\description{
This function plots the smooth triggering function for each component,
g(t), h(s), and k(m) for time, space, and marks, respectively.
}
